% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_lik.R
\name{all_model_info}
\alias{all_model_info}
\title{Synthesize model info}
\usage{
all_model_info(ts_result_list)
}
\arguments{
\item{ts_result_list}{All ts results}
}
\value{
joint df of model info
}
\description{
Synthesize model info
}
