% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ldats_plans.R
\name{build_ldats_analyses_plan}
\alias{build_ldats_analyses_plan}
\title{Build a drake plan for LDATS analysis}
\usage{
build_ldats_analyses_plan(datasets, max_topics = c(3), nseeds = 4,
  nchangepoints = c(2, 3), formulas = c("time", "intercept"),
  lda_control = list(), ts_control = list())
}
\arguments{
\item{datasets}{the drake plan for the datasets to analyze}

\item{max_topics}{the maximum number of topics to try}

\item{nseeds}{Number of seeds (replicate starts) to use for each 
value of \code{topics}. Must be conformable to \code{integer} value.}

\item{nchangepoints}{Vector of integers: which numbers of changepoints to try.}

\item{formulas}{Defaults to NULL. If nothing supplied, ~timename. If supplied, uses that.}
}
\value{
a drake plan
}
\description{
Construct the expanded Drake plan for LDA time series analyses
}
