% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ldats_plans.R
\name{build_ts_analysis_plan}
\alias{build_ts_analysis_plan}
\title{Build a drake plan for TS analysis}
\usage{
build_ts_analysis_plan(methods, datasets, lda_targets)
}
\arguments{
\item{datasets}{}
}
\value{
a drake plan
}
\description{
Build a drake plan for TS analysis
}
