% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ldats_plans.R
\name{build_ts_plan}
\alias{build_ts_plan}
\title{Build just a TS plan}
\usage{
build_ts_plan(ldamodels_names, nchangepoints = c(2, 3),
  formulas = c("time", "intercept"), ts_control = list())
}
\arguments{
\item{ldamodels_names}{Names of LDA models}

\item{nchangepoints}{n cpts}

\item{formulas}{formulas}

\item{ts_control}{passed to run_TS}
}
\value{
drake plan for TS section
}
\description{
Build just a TS plan
}
