% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ldats_plans.R
\name{build_ts_select_plan}
\alias{build_ts_select_plan}
\title{Build a drake plan for select_TS analysis}
\usage{
build_ts_select_plan(methods, cpt_targets)
}
\arguments{
\item{changepoint_models}{}
}
\value{
a drake plan
}
\description{
Build a drake plan for select_TS analysis
}
