% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ts_data.R
\name{check_cpt_format}
\alias{check_cpt_format}
\title{Check changepoint model format for TS select}
\usage{
check_cpt_format(cpt_model)
}
\arguments{
\item{data}{model to check}
}
\value{
TRUE or FALSE
}
\description{
Check whether the changepoint model is a changepoint model
}
