% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ts_data.R
\name{check_time_data_format}
\alias{check_time_data_format}
\title{Check data format for TS anaylysis}
\usage{
check_time_data_format(data)
}
\arguments{
\item{data}{dataset to check}
}
\value{
TRUE or FALSE
}
\description{
Check whether the timename column exists & is integers
}
