% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{collect_lda_result_summary}
\alias{collect_lda_result_summary}
\title{Make LDA result summary table}
\usage{
collect_lda_result_summary(lda_results)
}
\arguments{
\item{lda_results}{lda_results}
}
\value{
table of lda name, ntopics, nterms (species), ntimesteps
}
\description{
Make LDA result summary table
}
