% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{collect_lda_ts_results}
\alias{collect_lda_ts_results}
\title{Collect LDA and TS summaries in one df}
\usage{
collect_lda_ts_results(lda_result_summary, ts_result_summary)
}
\arguments{
\item{lda_result_summary}{lda_result_summary}

\item{ts_result_summary}{ts_result_summary}
}
\value{
df of tables joined
}
\description{
Collect LDA and TS summaries in one df
}
