% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{collect_ts_result_models_summary}
\alias{collect_ts_result_models_summary}
\title{Make detailed TS result summary table}
\usage{
collect_ts_result_models_summary(ts_results)
}
\arguments{
\item{ts_results}{All ts results (not selected)}
}
\value{
table of ts name, nchangepoints, formula, and AIC for each model
}
\description{
Make detailed TS result summary table
}
