% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{collect_ts_result_summary}
\alias{collect_ts_result_summary}
\title{Make TS result summary table}
\usage{
collect_ts_result_summary(selected_ts_results)
}
\arguments{
\item{ts_results}{ts_results}
}
\value{
table of ts name, nchangepoints
}
\description{
Make TS result summary table
}
