% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{expand_full_lik_results}
\alias{expand_full_lik_results}
\title{Expand full like}
\usage{
expand_full_lik_results(full_lik)
}
\arguments{
\item{result}{of get_full_lik}
}
\value{
model info with every AICc on its own line
}
\description{
Expand full like
}
