% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extract_lda_variable}
\alias{extract_lda_variable}
\title{Extract results from LDA results}
\usage{
extract_lda_variable(lda_result, variable_name)
}
\arguments{
\item{lda_result}{an LDA model}

\item{variable_name}{"ntopics", "ntimeseries", or "ntimesteps"}
}
\value{
value of that variable
}
\description{
Extract results from LDA results
}
