% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extract_max_topics}
\alias{extract_max_topics}
\title{Get max topics}
\usage{
extract_max_topics(lda_name)
}
\arguments{
\item{lda_name}{lda name}
}
\value{
limit on number of topics
}
\description{
Get max topics
}
