% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extract_ts_variable}
\alias{extract_ts_variable}
\title{Extract results from TS results}
\usage{
extract_ts_variable(ts_result, variable_name)
}
\arguments{
\item{ts_result}{a TS model}

\item{variable_name}{"nchangepoints", to be expanded}
}
\value{
value of that variable
}
\description{
Extract results from TS results
}
