% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{generalize_formula}
\alias{generalize_formula}
\title{Get generalized formula for TS model}
\usage{
generalize_formula(model_formula)
}
\arguments{
\item{model_formula}{string to process}
}
\value{
formula as "gamma ~ time" instead of "gamma ~ TIMENAME"
}
\description{
Get generalized formula for TS model
}
