% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_aicc}
\alias{get_aicc}
\title{Calculate AICc of observed data given estimates of beta and theta}
\usage{
get_aicc(beta_matrix, theta_matrix, lda_model, ts_model, counts_matrix)
}
\arguments{
\item{beta_matrix}{topic-term probabilities}

\item{theta_matrix}{document-topic probabilties}

\item{lda_model}{lda model, for df}

\item{ts_model}{ts model, for df}

\item{counts_matrix}{observed abundance data}
}
\value{
AICc
}
\description{
Calculate AICc of observed data given estimates of beta and theta
}
