% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_all_aiccs}
\alias{get_all_aiccs}
\title{Get all AICcs
Wrapper for \code{get_aicc}.}
\usage{
get_all_aiccs(bt_list, ldamodel, tsmodel, counts_matrix)
}
\arguments{
\item{bt_list}{list of beta and theta values}

\item{ldamodel}{source lda model}

\item{tsmodel}{source ts model}

\item{counts_matrix}{abundance}
}
\value{
list of AICcs from each estimate of theta
}
\description{
Get all AICcs
Wrapper for \code{get_aicc}.
}
