% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_bt}
\alias{get_bt}
\title{#' Generate table of LDA and TS info
#'
#' @param lda_model_names names of LDA models
#' @param ts_model_names names of TS models
#'
#' @return df of LDA nad TS model info
#' @export
#' @importFrom dplyr left_join
make_lda_ts_table <- function(lda_model_names, ts_model_names) {
ts_table <- make_ts_table(ts_model_names)
lda_table <- make_lda_table(lda_model_names)\preformatted{lda_ts_table <- dplyr::left_join(ts_table, lda_table, by = c("seed", "k"))
return(lda_ts_table)
}

}
Get beta and theta matrices from lda and ts model}
\usage{
get_bt(lda_model, ts_model)
}
\arguments{
\item{lda_model}{LDA model}

\item{ts_model}{TS model fit using that LDA model}
}
\value{
list of beta_vals and all theta estimats from the TS model
}
\description{
#' Generate table of LDA and TS info
#'
#' @param lda_model_names names of LDA models
#' @param ts_model_names names of TS models
#'
#' @return df of LDA nad TS model info
#' @export
#' @importFrom dplyr left_join
make_lda_ts_table <- function(lda_model_names, ts_model_names) {
ts_table <- make_ts_table(ts_model_names)
lda_table <- make_lda_table(lda_model_names)\preformatted{lda_ts_table <- dplyr::left_join(ts_table, lda_table, by = c("seed", "k"))
return(lda_ts_table)
}

}
Get beta and theta matrices from lda and ts model
}
