% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{get_data_names}
\alias{get_data_names}
\title{Get data name out of model name}
\usage{
get_data_names(model_name, model_type = "ts_notselected")
}
\arguments{
\item{model_name}{model name, string to process}

\item{model_type}{"ts_notselected", "ts_selected", "lda_selected"}
}
\value{
data name
}
\description{
Get data name out of model name
}
