% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{get_filtered}
\alias{get_filtered}
\title{Get whether a model was run on filtered data
From model name}
\usage{
get_filtered(model_name)
}
\arguments{
\item{model_name}{string to process}
}
\value{
"filtered", "complete"
}
\description{
Get whether a model was run on filtered data
From model name
}
