% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_full_lik}
\alias{get_full_lik}
\title{Get document-term probabilities and AICc from ts_model}
\usage{
get_full_lik(ts_list)
}
\arguments{
\item{ts_list}{result of run_TS}
}
\value{
list of data, lda model, ts model, betas & thetas, and AICc estimates
}
\description{
Get document-term probabilities and AICc from ts_model
}
