% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_loglik}
\alias{get_loglik}
\title{Calculate loglik of observed data given estimates of beta and theta}
\usage{
get_loglik(beta_matrix, theta_matrix, counts_matrix)
}
\arguments{
\item{beta_matrix}{topic-term probabilities}

\item{theta_matrix}{document-topic probabilties}

\item{counts_matrix}{observed abundance data}
}
\value{
loglik
}
\description{
Calculate loglik of observed data given estimates of beta and theta
}
