% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_data_processing.R
\name{get_portal_annual_data}
\alias{get_portal_annual_data}
\title{Get portal annual data}
\usage{
get_portal_annual_data()
}
\value{
dataset of portal annual data
}
\description{
Get portal annual data
}
