% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_relevant_covars}
\alias{get_relevant_covars}
\title{Extract the covariates used in a formula}
\usage{
get_relevant_covars(covariates_table, ts_formula)
}
\arguments{
\item{covariates_table}{from a data object}

\item{ts_formula}{formula for a ts model}
}
\value{
the subset of columns in the covariates table used in the formula
}
\description{
Extract the covariates used in a formula
}
