% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_theta}
\alias{get_theta}
\title{Get theta from a ts model}
\usage{
get_theta(ts_model, sim = 1)
}
\arguments{
\item{ts_model}{TS_on_LDA}

\item{sim}{which draw from the posterior}
}
\value{
Theta matrix for that draw
}
\description{
The thetas are the document-topic probabilities for each draw from the posterior for Eta and rho.
}
