% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ldats_plans.R
\name{logLik.LDA_VEM}
\alias{logLik.LDA_VEM}
\title{Logliklihood (From updated LDATS)}
\usage{
\method{logLik}{LDA_VEM}(object, ...)
}
\arguments{
\item{object}{for example an LDA model}

\item{...}{other arugments}
}
\value{
loglikelihood
}
\description{
Logliklihood (From updated LDATS)
}
