% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ldats_plans.R
\name{logLik}
\alias{logLik}
\title{Logliklihood (From updated LDATS)}
\usage{
logLik(object, ...)
}
\arguments{
\item{object}{for example an LDA model}

\item{...}{other arugments}
}
\value{
loglikelihood
}
\description{
Logliklihood (From updated LDATS)
}
