% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{make_lda_table}
\alias{make_lda_table}
\title{Generate table of LDA model info}
\usage{
make_lda_table(lda_model_names)
}
\arguments{
\item{lda_model_names}{names of an LDA_set}
}
\value{
dataframe of name, index, seed, and k
}
\description{
Generate table of LDA model info
}
