% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{make_ts_table}
\alias{make_ts_table}
\title{Generate table of TS model info}
\usage{
make_ts_table(ts_model_names)
}
\arguments{
\item{ts_model_names}{names of a set of TS models}
}
\value{
data frame of corresponding TS model name, index, k, seed, nchangepoints, formula
}
\description{
Generate table of TS model info
}
