% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_abundances}
\alias{predict_abundances}
\title{Predict abundances given model}
\usage{
predict_abundances(full_lik, seed = 1977)
}
\arguments{
\item{full_lik}{result of get_full_lik}

\item{seed}{for reprod}
}
\value{
list of predictions
}
\description{
Predict abundances given model
}
