% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_data_processing.R
\name{remove_transients}
\alias{remove_transients}
\title{Remove transient species}
\usage{
remove_transients(dataset, threshold = 0.3)
}
\arguments{
\item{threshold}{Cutoff occurrence rate}

\item{document_term_table}{}
}
\value{
dataset with modified document term table
}
\description{
Remove species that occur in < X% of samples.
}
\details{
Values for transients could be .1-.35
}
