% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_LDA.R
\name{run_LDA}
\alias{run_LDA}
\title{Run Latent Dirichlet Allocation on Tabular Data}
\usage{
run_LDA(data, max_topics = 6, nseeds = 200, control = list())
}
\arguments{
\item{data}{a data.frame or tibble; each row is an observation (e.g. in
time or space), and each column is a variable. Here, the common usage is
for each column to be a species or taxon, and each row to be an observed
sample. In the original specification for LDA, each row is a document,
and each column is a word, with the entries being the counts of the words
in each document.}

\item{max_topics}{the maximum number of topics to try (the function will
test a number of topics from 2 to \code{max_topics})}

\item{nseeds}{Number of seeds (replicate starts) to use for each 
value of \code{topics}. Must be conformable to \code{integer} value.}

\item{control}{A \code{list} of parameters to control the running and 
selecting of LDA models. Values not input assume default values set 
by \code{\link{LDA_set_control}}. Values for running the LDAs replace 
defaults in (\code{LDAcontol}, see \code{\link[topicmodels]{LDA}} (but if
 \code{seed} is given, it will be overwritten; use \code{iseed} instead).}
}
\value{
list of \code{lda = [the set of LDA models]} from running LDATS::parLDA()\code{and}upstream = list(data = data)`
}
\description{
Test the Latent Dirichlet Allocation (LDA) model on the data
with different number of topics (from 2 to \code{max_topics}), select the best
one using AIC, and return the model object that is selected.
}
