% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_TS.R
\name{run_TS}
\alias{run_TS}
\title{Run Changepoint Model on LDA + TimeSeries Data}
\usage{
run_TS(data, ldamodels, formulas = ~1, nchangepoints = c(0:6),
  weighting = "proportional", control = LDATS::TS_controls_list())
}
\arguments{
\item{data}{The dataset, in the \code{MATSS} format. Including covariates & metadata.}

\item{ldamodels}{The LDA model(s) produced by running \code{run_LDA} on the dataset.}

\item{formulas}{Defaults to ~1.}

\item{nchangepoints}{Vector of integers: which numbers of changepoints to try.}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}

\item{weights}{Weight samples equally or proportional to number of individuals captured. Defaults to \code{'proprotional'}; any other value will weight samples equally.}
}
\value{
the changepoint model(s) resulting from running \code{LDATS::TS_on_LDA}.
}
\description{
Run the timeseries model from \code{ldats} on LDA models and time-series data.
}
