% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_LDA.R
\name{select_LDA_list}
\alias{select_LDA_list}
\title{LDA select}
\usage{
select_LDA_list(lda_list)
}
\arguments{
\item{lda_list}{list of LDA + data}
}
\value{
list of selected LDA + data
}
\description{
LDA select
}
