% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_TS.R
\name{select_TS_list}
\alias{select_TS_list}
\title{Select TS from list}
\usage{
select_TS_list(ts_list)
}
\arguments{
\item{ts_list}{list of ts, upstream}
}
\value{
list of selected ts, upstream
}
\description{
Select TS from list
}
