% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_lik.R
\name{subset_data}
\alias{subset_data}
\title{Subset data into training & test sets}
\usage{
subset_data(data, n_segs = NULL, sequential = T, buffer = NULL,
  which_seg = NULL)
}
\arguments{
\item{data}{Beginning data}

\item{n_segs}{nb of segments}

\item{sequential}{sequential or not - to be added}

\item{buffer}{buffer or not - to be added}

\item{which_seg}{which segment to return - for drake}
}
\value{
list of subsetted data, covariates, metadata, test data, test covs, OR list of all above lists
}
\description{
Subset data into training & test sets
}
