% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_lik.R
\name{ts_AICc}
\alias{ts_AICc}
\title{All AICc for a single TS model}
\usage{
ts_AICc(ts_model, lda_model, data)
}
\arguments{
\item{ts_model}{one TS fit}

\item{lda_model}{upstream LDA model}

\item{data}{data list}
}
\value{
vector of all AICc
}
\description{
All AICc for a single TS model
}
