% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_lik.R
\name{ts_predict}
\alias{ts_predict}
\title{Generate predictions from ts model}
\usage{
ts_predict(l = NULL, lda_model = NULL, ts_model = NULL,
  data = NULL, sim = NULL, seed = 1977)
}
\arguments{
\item{l}{row from a model_info table as a list}

\item{lda_model}{or you can give lda model}

\item{ts_model}{and ts model}

\item{data}{and data}

\item{sim}{sim to pull?}

\item{seed}{seed to set?}

\item{cache_location}{cache location}
}
\value{
matrix of predicted and observed data
}
\description{
Generate predictions from ts model
}
