% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{ts_predict_corpus}
\alias{ts_predict_corpus}
\title{Predict abundances for new data based on TS model}
\usage{
ts_predict_corpus(ts_model, lda_model, data, predict_data = NULL,
  draw = 1)
}
\arguments{
\item{ts_model}{one TS fit}

\item{lda_model}{the upstream LDA model}

\item{data}{base data}

\item{predict_data}{list of abundance and covariates for timesteps to predict over}

\item{draw}{which draw from the posterior}
}
\value{
sampled corpus
}
\description{
Predict abundances for new data based on TS model
}
