% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_lik.R
\name{ts_test_loglik}
\alias{ts_test_loglik}
\title{All loglik for a single TS model vs. hold-out data}
\usage{
ts_test_loglik(ts_model, lda_model, data)
}
\arguments{
\item{ts_model}{one TS fit}

\item{lda_model}{upstream LDA model}

\item{data}{data list}
}
\value{
vector of all logliks
}
\description{
All loglik for a single TS model vs. hold-out data
}
