% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ldats_plans.R
\name{build_ldats_analyses_plan}
\alias{build_ldats_analyses_plan}
\title{Build a drake plan for LDATS analysis}
\usage{
build_ldats_analyses_plan(datasets, max_topics = 3, nseeds = 4,
  nchangepoints = c(2, 3))
}
\arguments{
\item{datasets}{the drake plan for the datasets to analyze}

\item{max_topics}{the maximum number of topics to try (the function will
test a number of topics from 2 to \code{max_topics})}

\item{nseeds}{Number of seeds (replicate starts) to use for each 
value of \code{topics}. Must be conformable to \code{integer} value.}

\item{nchangepoints}{Vector of integers: which numbers of changepoints to try.}
}
\value{
a drake plan
}
\description{
Construct the expanded Drake plan for LDA time series analyses
}
