% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ldats_plans.R
\name{build_ts_analysis_plan}
\alias{build_ts_analysis_plan}
\title{Build a drake plan for TS analysis}
\usage{
build_ts_analysis_plan(methods, datasets, lda_targets)
}
\arguments{
\item{methods}{the drake plan for the methods to run}

\item{datasets}{the drake plan for the datasets to analyze}

\item{lda_targets}{the drake plan for the LDA targets}
}
\value{
a drake plan
}
\description{
Construct the expanded Drake plan for LDA time series analyses
}
