% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ldats_plans.R
\name{build_ts_select_plan}
\alias{build_ts_select_plan}
\title{Build a drake plan for select_TS analysis}
\usage{
build_ts_select_plan(methods, cpt_targets)
}
\arguments{
\item{methods}{the drake plan for the methods to run}

\item{cpt_targets}{the drake plan for the changepoint models to construct}
}
\value{
a drake plan
}
\description{
Construct the expanded Drake plan for selecting changepoint models
}
