% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_doc_lik}
\alias{get_doc_lik}
\title{Compute loglikelihood of observed data given probabilities}
\usage{
get_doc_lik(counts_matrix, p_matrix, index)
}
\arguments{
\item{counts_matrix}{observed abundance data}

\item{p_matrix}{predicted abundance probabilities}

\item{index}{which document; for iterating over all documents}
}
\value{
loglik of a single document
}
\description{
Compute loglikelihood of observed data given probabilities
}
