% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_TS.R
\name{run_TS}
\alias{run_TS}
\title{Run Changepoint Model on LDA + TimeSeries Data}
\usage{
run_TS(data, ldamodels, formulas = ~1, nchangepoints = c(0:6),
  weighting = "proportional", control = LDATS::TS_controls_list())
}
\arguments{
\item{formulas}{Vector of \code{\link[stats]{formula}}(s) for the 
continuous (non-change point) component of the time series models. Any 
predictor variable included in a formula must also be a column in the
\code{document_covariate_table}. Each element (formula) in the vector
is evaluated for each number of change points and each LDA model.}

\item{nchangepoints}{Vector of \code{integer}s corresponding to the number 
of change points to include in the time series models. 0 is a valid input 
corresponding to no change points (\emph{i.e.}, a singular time series
model), and the current implementation can reasonably include up to 6 
change points. Each element in the vector is the number of change points 
used to segement the data for each formula (entry in \code{formulas}) 
component of the TS model, for each selected LDA model.}

\item{control}{Class \code{TS_controls} list, holding control parameters
for the Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls, generated by 
\code{\link{TS_controls_list}}.}

\item{max_topics}{the maximum number of topics to try (the function will
test a number of topics from 2 to \code{max_topics})}
}
\value{
the best fit model object, from running \code{LDATS::parLDA()}
}
\description{
Run the timeseries model from \code{ldats} on LDA models and time-series data.
}
