% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_TS.R
\name{run_TS}
\alias{run_TS}
\title{Run Changepoint Model on LDA + TimeSeries Data}
\usage{
run_TS(data, ldamodels, formulas = c("intercept", "time"),
  nchangepoints = 0:6, weighting = "proportional", control = list(nit
  = 1000, magnitude = max(1, floor(0.03 * nrow(data$abundance)))))
}
\arguments{
\item{data}{The dataset, in the \code{MATSS} format. Including covariates & metadata.}

\item{ldamodels}{The LDA model(s) produced by running \code{run_LDA} on the dataset.}

\item{formulas}{Defaults to NULL. If nothing supplied, ~timename. If supplied, uses that.}

\item{nchangepoints}{Vector of integers: which numbers of changepoints to try.}

\item{weighting}{Weight samples equally or proportional to number of individuals captured. Defaults to \code{'proprotional'}; any other value will weight samples equally.}

\item{control}{A \code{list} of parameters to control the fitting of the
Time Series model including the parallel tempering Markov Chain 
Monte Carlo (ptMCMC) controls. Values not input assume defaults set by 
\code{\link{TS_control}}.}
}
\value{
list of \code{ts = [the changepoint model(s)]} and \code{upstream = list(data = data, lda = lda)} resulting from running \code{LDATS::TS_on_LDA}.
}
\description{
Run the timeseries model from \code{ldats} on LDA models and time-series data.
}
