% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{sample_corpus}
\alias{sample_corpus}
\title{Sample a corpus given sample sizes and ps of documents}
\usage{
sample_corpus(docterm_ps, obs_dat)
}
\arguments{
\item{docterm_ps}{probabilities of terms in each document}

\item{sample_sizes}{sizes of documents}
}
\value{
sampled corpus
}
\description{
Sample a corpus given sample sizes and ps of documents
}
