% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_TS.R
\name{select_TS}
\alias{select_TS}
\title{Run Changepoint Model on LDA + TimeSeries Data}
\usage{
select_TS(changepoint_models)
}
\value{
best fit TS model
}
\description{
Run the timeseries model from \code{ldats} on LDA models and time-series data.
}
